quick commit

+++ routes/api.php
@@ -289,6 +289,6 @@ Route::middleware('auth:sanctum')->group(function () {
     Route::get('/operationinfo/{id}', 'NewOperationsController@getInfo');
     Route::put('/newcampaignusersremove/{id}/{opID}', 'OperationUserController@updateremove');
     Route::put('/newcampaignusersadd/{id}/{opID}', 'OperationUserController@updateadd');
-    Route::post('/newcampaignusers/{opID}', 'OperationUserController@store');
-    Route::delete('/newcampaignusers/{id}/{opID}', 'OperationUserController@destroy');
+    Route::post('/newcampaignusers/{opID}/{userid}', 'OperationUserController@store');
+    Route::delete('/newcampaignusers/{id}/{opID}/{userid}', 'OperationUserController@destroy');
 });
